moved mapper to seperate file for simplicity
