+++ configs-datasets/agera5_monthly/agera5_monthly.py
@@ -28,7 +28,7 @@ from stacbuilder import (
     FileCollectorConfig,
 )
 from stacbuilder.boundingbox import BoundingBox
-from stacbuilder.collector import IMetadataCollector, MetadataFromFileCollector
+from stacbuilder.collector import IMetadataCollector, MetadataCollector
 
 
 def build_collection(
@@ -44,7 +44,7 @@ def build_collection(
     if output_dir and not isinstance(output_dir, Path):
         output_dir = Path(output_dir).expanduser().absolute()
 
-    collector = MetadataFromFileCollector.from_config(collection_config=coll_cfg, file_coll_cfg=file_coll_cfg)
+    collector = MetadataCollector.from_config(collection_config=coll_cfg, file_coll_cfg=file_coll_cfg)
 
     if output_dir and not isinstance(output_dir, Path):
         output_dir = Path(output_dir).expanduser().absolute()
