+++ configs-datasets/evoland/spatial_features/build.py
@@ -13,7 +13,7 @@ from stacbuilder import (
     validate_collection,
 )
 from stacbuilder.builder import AssetMetadataPipeline
-from stacbuilder.collector import MetadataFromFileCollector
+from stacbuilder.collector import MetadataCollector
 from stacbuilder.config import CollectionConfig, FileCollectorConfig
 from stacbuilder.stacapi import (
     AuthSettings,
@@ -48,7 +48,7 @@ if output_path and not isinstance(output_path, Path):
     output_path = Path(output_path).expanduser().absolute()
 
 # Define collector
-collector = MetadataFromFileCollector.from_config(collection_config=coll_cfg, file_coll_cfg=file_coll_cfg)
+collector = MetadataCollector.from_config(collection_config=coll_cfg, file_coll_cfg=file_coll_cfg)
 
 # create pipeline
 pipeline: AssetMetadataPipeline = AssetMetadataPipeline.from_config(
